框架搭建

+++ msgcenter/apimap.go
@@ -2,7 +2,7 @@ package msgcenter
 
 import "zjh/Test/apitest"
 
-type HandleFunc func(uint64, []byte) interface{}
+type HandleFunc func(uint32, []byte) interface{}
 
 type APIInfo struct {
 	APIFunction HandleFunc
@@ -12,6 +12,6 @@ const (
 	TestFunc = 0
 )
 
-var ApiMap = map[uint16]*APIInfo{
+var ApiMap = map[uint32]*APIInfo{
 	TestFunc: &APIInfo{APIFunction: apitest.TestFunc},
 }
